Add Error monad to allow for conversion to fail
Replace calls to `error` with proper error handling. Should fix problems
with Accelerate program being partially printed before an error occurs.

Note that the dependency on mtl has been increased to 2.2.1, because of
the use of `Control.Monad.Except` which is a replacement for the
deprecated `Control.Monad.Error` module.

Closes #1